Added exported enum for return tuple index positions.
This is handy, especially when one wants to use just the publish and not necessarily the state. In Typescript, you'll get unused variable errors if you don't use the state variable and using underscore names opens another can of works.

The preferred pattern to solve this is:

const publish = usePubSub<MyType>('myTopic', myDefaultValue)[TupleIndex.Publish]